+++ solutions/minCostClimbingStairs.cpp
@@ -12,7 +12,7 @@ public:
         min_cost_per_stair[0] = 0;
         min_cost_per_stair[1] = 0;
 
-        for (int i = 2; i <= cost.size(); i++)
+        for (size_t i = 2; i <= cost.size(); i++)
         {
             min_cost_per_stair[i] = min(min_cost_per_stair[i - 1] + cost[i - 1], min_cost_per_stair[i - 2] + cost[i - 2]);
         }
